+++ packages/core-js/full/weak-map/get-or-insert.js
@@ -1,6 +1,4 @@
 'use strict';
-require('../../modules/es.weak-map');
-require('../../modules/esnext.weak-map.get-or-insert');
-var entryUnbind = require('../../internals/entry-unbind');
+var parent = require('../../actual/weak-map/get-or-insert');
 
-module.exports = entryUnbind('WeakMap', 'getOrInsert');
+module.exports = parent;