+++ packages/core-js/stage/2.7.js
@@ -1,6 +1,6 @@
 'use strict';
 var parent = require('./3');
 
-require('../proposals/map-upsert-v4');
+require('../proposals/joint-iteration');
 
 module.exports = parent;