+++ packages/core-js/stage/2.js
@@ -14,7 +14,6 @@ require('../proposals/async-explicit-resource-management');
 require('../proposals/decorators');
 require('../proposals/decorator-metadata');
 require('../proposals/iterator-helpers');
-require('../proposals/joint-iteration');
 require('../proposals/map-upsert-stage-2');
 require('../proposals/math-clamp-v2');
 require('../proposals/set-methods');