+++ tests/unit-pure/esnext.map.get-or-insert-computed.js
@@ -1,5 +1,5 @@
 import { STRICT } from '../helpers/constants.js';
-import Map from 'core-js-pure/full/map';
+import Map from 'core-js-pure/actual/map';
 import from from 'core-js-pure/es/array/from';
 
 QUnit.test('Map#getOrInsertComputed', assert => {