+++ tests/unit-pure/esnext.map.get-or-insert.js
@@ -1,4 +1,4 @@
-import Map from 'core-js-pure/full/map';
+import Map from 'core-js-pure/actual/map';
 import from from 'core-js-pure/es/array/from';
 
 QUnit.test('Map#getOrInsert', assert => {